providers: rename noop to metal
We don't generally use the noop provider for other purposes, so switch to
a clearer name.
